+++ packages/hspec/CHANGELOG.md
@@ -6,7 +6,18 @@ and this project adheres to [PVP](https://pvp.haskell.org/).
 
 ## [Unreleased]
 
-Nothing
+### Fixed
+
+* Added all defaults for now the FromConfig instance actually works
+
+### Changed
+
+* `"checks"` is a valid value for `formatter` key (with `hspec >= 2.7.10`)
+* `failed-examples` is a valid value for `formatter`
+
+### Added
+
+* `randomize` key is used for configuring `configRandomize`
 
 ## [v1.1.0.0] 2021-03-07